As noticed by a PixelLight user, the data type returned by "System::GetMicroseconds()" was "uint32" which is critical for the stored data and "unsigned long" is commonly used instead. So, methods like "System::GetMicroseconds()" are now using "uint64" instead of "uint32".

+++ PLBase/PLGeneral/include/PLGeneral/System/Mutex.h
@@ -108,7 +108,7 @@ class Mutex {
 		*  @see
 		*    - Lock()
 		*/
-		PLGENERAL_API bool TryLock(uint32 nTimeout);
+		PLGENERAL_API bool TryLock(uint64 nTimeout);
 
 		/**
 		*  @brief
