+++ PLBase/PLGeneral/src/System/Mutex.cpp
@@ -85,7 +85,7 @@ bool Mutex::Lock()
 *  @brief
 *    Locks the mutex, but only wait until timeout
 */
-bool Mutex::TryLock(uint32 nTimeout)
+bool Mutex::TryLock(uint64 nTimeout)
 {
 	// Call system implementation
 	return m_pMutexImpl->TryLock(nTimeout);
